+++ expandedstorage/changelog.md
@@ -1,1 +1,4 @@
-- Fixed shipping some debug code.
+- Added support for inventory sorter's buttons ( moved mine when mod is loaded )
+- Fixed narration of next and previous page buttons.
+- Made tooltips follow cursor in scren type select screen.
+- Added tooltip to select screen type button ( settings button )